Some fixes
I did not compile it. It may need to fix the references to compile. 

Please proof read. Also please check if OQS OpenSSH encodes K_PQ as mpint. Since it sia 256-bit value it could, it is not necessary to encode it as as a byte arraw because it is not the concatenation of two things. 

I would also suggest to consider dropping MLKEM-512, but since you have the IANA assignments, I guess you can keep it.